Add files via upload
Create an array of strings and a separate single string. Test to see if the strings in the array begin and end with the characters in the single string. For each string that does, extract the substring between the beginning and ending characters and add that substring to a second array.

For example, if the single string is "_a1_", and one of the strings in the array is "_a1_testing_a1_", the substring "testing" would be added to the second array.